backend/requirements.txt
fastapi
uvicorn
openai
requests
python-dotenv
beautifulsoup4
langchain